build(webpack): convert FeatureFlaggedSriPlugin to TypeScript
